chore: fix code climate issues
